+++ creator-node/src/index.ts
@@ -20,7 +20,10 @@ import DBManager from './dbManager'
 
 import { logger } from './logging'
 import { sequelize } from './models'
-import DiskManager from './diskManager'
+import {
+  emptyTmpTrackUploadArtifacts,
+  sweepSubdirectoriesInFiles
+} from './diskManager'
 
 const EthereumWallet = require('ethereumjs-wallet')
 const redisClient = require('./redis')
@@ -111,7 +114,7 @@ const startAppForPrimary = async () => {
   await setupDbAndRedis()
 
   const startTime = Date.now()
-  const size = await DiskManager.emptyTmpTrackUploadArtifacts()
+  const size = await emptyTmpTrackUploadArtifacts()
   logger.info(
     `old tmp track artifacts deleted : ${size} : ${
       (Date.now() - startTime) / 1000
@@ -179,7 +182,7 @@ const startAppForPrimary = async () => {
   // wait one minute before starting this because it might cause init to degrade
   if (config.get('diskPruneEnabled')) {
     setTimeout(() => {
-      DiskManager.sweepSubdirectoriesInFiles()
+      sweepSubdirectoriesInFiles()
     }, 60_000)
   }
 }
